renamed useAllowedNetwork; updated imports
