+++ packages/nextjs/hooks/scaffold-eth/index.ts
@@ -14,3 +14,4 @@ export * from "./useScaffoldWriteContract";
 export * from "./useTargetNetwork";
 export * from "./useTransactor";
 export * from "./useWatchBalance";
+export * from "./useAllowedNetwork";
