+++ packages/nextjs/hooks/scaffold-eth/useNetworkColor.ts
@@ -1,6 +1,6 @@
-import { useAllowedNetwork } from "./useAllowedChain";
 import { useTargetNetwork } from "./useTargetNetwork";
 import { useTheme } from "next-themes";
+import { useAllowedNetwork } from "~~/hooks/scaffold-eth";
 import { AllowedChainIds, ChainWithAttributes } from "~~/utils/scaffold-eth";
 
 export const DEFAULT_NETWORK_COLOR: [string, string] = ["#666666", "#bbbbbb"];
