+++ packages/nextjs/hooks/scaffold-eth/useScaffoldReadContract.ts
@@ -1,9 +1,9 @@
 import { useEffect } from "react";
-import { useAllowedNetwork } from "./useAllowedChain";
 import { QueryObserverResult, RefetchOptions, useQueryClient } from "@tanstack/react-query";
 import type { ExtractAbiFunctionNames } from "abitype";
 import { ReadContractErrorType } from "viem";
 import { useBlockNumber, useReadContract } from "wagmi";
+import { useAllowedNetwork } from "~~/hooks/scaffold-eth";
 import { useDeployedContractInfo } from "~~/hooks/scaffold-eth";
 import { AllowedChainIds } from "~~/utils/scaffold-eth";
 import {
